FE-51 Refactor : modules내 스타일컴포넌트 분리 1. 스타일 컴포넌트 분리했습니다 2. import순서 정리했습니다.
