+++ src/components/modules/Comment/Comment.jsx
@@ -1,48 +1,9 @@
-import styled from 'styled-components';
+import { Link } from 'react-router-dom';
+import { useRecoilValue } from 'recoil';
 import Profile from '../../atoms/Profile/Profile';
 import Icon from '../../atoms/Icon/Icon';
 import { accountnameValue } from '../../../atoms';
-import { useRecoilValue } from 'recoil';
-import { Link } from 'react-router-dom';
-
-const CommentUserInfoLi = styled.li`
-  display: flex;
-  gap: 12px;
-  justify-content: space-between;
-`;
-const CommentTxt = styled.div`
-  width: 100%;
-`;
-
-const UserInfo = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 6px;
-`;
-const Time = styled.span`
-  flex-grow: 1;
-  font-weight: 400;
-  font-size: 10px;
-  line-height: 13px;
-  color: ${props => props.theme.color.gray.d4};
-`;
-
-const UserId = styled.strong`
-  font-weight: 500;
-  font-size: 14px;
-  line-height: 18px;
-`;
-
-const CommentContent = styled.p`
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 18px;
-  margin-top: 16px;
-  padding-right: 10px;
-  color: #333333;
-  word-break: keep-all;
-  word-wrap: break-word;
-`;
+import * as S from './style';
 
 const Comment = ({
   data,
@@ -75,7 +36,7 @@ const Comment = ({
   };
 
   return (
-    <CommentUserInfoLi>
+    <S.CommentUserInfoLi>
       <Link to={`/yourProfile/${data.author.accountname}`}>
         <Profile
           size='42px'
@@ -84,12 +45,12 @@ const Comment = ({
           borderRadius={props => props.theme.borderRadius.circle}
         />
       </Link>
-      <CommentTxt>
-        <UserInfo>
+      <S.CommentTxt>
+        <S.UserInfo>
           <Link to={`/yourProfile/${data.author.accountname}`}>
-            <UserId>{data.author.username}</UserId>
+            <S.UserId>{data.author.username}</S.UserId>
           </Link>
-          <Time>{timeMsg}</Time>
+          <S.Time>{timeMsg}</S.Time>
           <Icon
             id={data.id}
             accountname={data.author.accountname}
@@ -99,10 +60,10 @@ const Comment = ({
             isMyComment={isMyComment}
             onClick={handleMoreBtn}
           />
-        </UserInfo>
-        <CommentContent>{data.content}</CommentContent>
-      </CommentTxt>
-    </CommentUserInfoLi>
+        </S.UserInfo>
+        <S.CommentContent>{data.content}</S.CommentContent>
+      </S.CommentTxt>
+    </S.CommentUserInfoLi>
   );
 };
