+++ src/components/modules/CustomFileInput/CustomFileInput.stories.js
@@ -1,11 +1,13 @@
-import CustomFileInput from './CustomFileInput';
+// import CustomFileInput from './CustomFileInput';
 
-export default {
-  title: 'Modules/CustomFileInput',
-  component: CustomFileInput,
-};
+// export default {
+//   title: 'Modules / CustomFileInput',
+//   component: CustomFileInput,
+// };
 
-const Template = args => <CustomFileInput {...args} />;
+// const Template = args => <CustomFileInput {...args} />;
 
-export const CustomFileInputStory = Template.bind({});
-CustomFileInputStory.args = {};
+// export const CustomFileInputStory = Template.bind({});
+// CustomFileInputStory.args = {
+
+// };
