+++ src/components/modules/CustomFileInput/style.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const CustomFileInputWrapper = styled.div`
+  text-align: center;
+  label {
+    position: relative;
+    cursor: pointer;
+  }
+  button {
+    position: absolute;
+    bottom: 3px;
+    right: 3px;
+  }
+`;
