+++ src/components/modules/DropUp/DropUp.stories.js
@@ -1,17 +1,17 @@
-import DropUp from './DropUp';
+// import DropUp from './DropUp';
 
-export default {
-  title: 'Modules / Modal',
-  component: DropUp,
-};
+// export default {
+//   title: 'Modules / DropUp',
+//   component: DropUp,
+// };
 
-const Template = args => <DropUp {...args} />;
+// const Template = args => <DropUp {...args} />;
 
-export const PostDropUpStory = Template.bind({});
-PostDropUpStory.args = {
-  menu: ['삭제', '수정', '웹사이트에서 상품 보기'],
-};
-export const ProfileDropUpStory = Template.bind({});
-ProfileDropUpStory.args = {
-  menu: ['설정 및 개인정보', '로그아웃'],
-};
+// export const PostDropUpStory = Template.bind({});
+// PostDropUpStory.args = {
+//   menu: ['삭제', '수정', '웹사이트에서 상품 보기'],
+// };
+// export const ProfileDropUpStory = Template.bind({});
+// ProfileDropUpStory.args = {
+//   menu: ['설정 및 개인정보', '로그아웃'],
+// };
