+++ src/components/modules/InputBox/InputBox.stories.js
@@ -1,3 +1,4 @@
+import React from 'react';
 import InputBox from './InputBox';
 
 export default {
@@ -10,6 +11,6 @@ const Template = args => <InputBox {...args} />;
 export const EmailInputBoxStory = Template.bind({});
 EmailInputBoxStory.args = {
   label: '이메일',
-  placeholder: '이메일 입력해주세염 뿌우',
+  placeholder: '이메일 입력해주세',
   type: 'text',
 };
