+++ src/components/modules/LikeBtn/style.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const CountNum = styled.dd`
+  font-size: 12px;
+  line-height: 12px;
+  font-weight: 400;
+  display: inline-block;
+  vertical-align: top;
+  margin-top: 5px;
+  color: ${props => props.theme.color.gray.d4};
+  margin-left: 6px;
+`;
