+++ src/components/modules/Post/Post.stories.js
@@ -1,10 +1,10 @@
-import Post from './Post';
+// import Post from './Post';
 
-export default {
-  title: 'Modules / Post',
-  component: Post,
-};
+// export default {
+//   title: 'Modules / Post',
+//   component: Post,
+// };
 
-const Template = args => <Post {...args} />;
+// const Template = args => <Post {...args} />;
 
-export const PostStory = Template.bind({});
+// export const PostStory = Template.bind({});
